misc/trivia/challenge/requirements.txt
discord.py
db-sqlite3